WIP: Improvement of the ORM (#10)
This PR is a major refactor of the Model and query system. It adds
[SQLModel](https://sqlmodel.tiangolo.com/tutorial/fastapi/limit-and-offset/?h=offset#add-a-limit-and-offset-to-the-query-parameters),
which is similar to our current Model, and includes a QueryBuilder on
top of it to provide an API closer to Active Record or Eloquent.

Here's an example of [grace's
thread](https://github.com/Code-Society-Lab/grace/blob/main/bot/models/extensions/thread.py)
model:
```python
# bot/models/extensions/thread.py
from grace.model import Model, Field, Column, Text, Integer


class Thread(Model):
    __tablename__ = 'threads'  # only needed because for old tables
    
    id: int | None = Field(default=None, primary_key=True)
    title: str
    content: str = Field(sa_type=Text)
    recurrence: Recurrence = Field(sa_column=Column(Integer))
```

```python
thread = Thread.create(
    title="Daily Standup",
    content="Long content here...",
    recurrence=Recurrence.DAILY
)

daily_threads = Thread.where(Thread.recurrence == Recurrence.DAILY).all()
thread_by_id = Thread.find(1)

thread.recurrence = Recurrence.WEEKLY
thread.save()
```

+++ grace/application.py
@@ -1,32 +1,22 @@
-from os import environ
 from configparser import SectionProxy
-
-from coloredlogs import install
 from logging import basicConfig, critical
 from logging.handlers import RotatingFileHandler
-
+from os import environ
+from pathlib import Path
 from types import ModuleType
-from typing import Generator, Any, Union, Dict, Optional, no_type_check
+from typing import Any, Dict, Generator, Optional, Union, no_type_check
 
-from sqlalchemy import create_engine
+from coloredlogs import install
 from sqlalchemy.engine import Engine
 from sqlalchemy.exc import OperationalError
-from sqlalchemy.orm import (
-    declarative_base,
-    sessionmaker,
-    Session,
-    DeclarativeMeta
-)
-from sqlalchemy_utils import (
-    database_exists,
-    create_database,
-    drop_database
-)
-from pathlib import Path
+from sqlalchemy.orm import DeclarativeMeta, declarative_base
+from sqlalchemy_utils import create_database, database_exists, drop_database
+from sqlmodel import Session, create_engine
+
 from grace.config import Config
 from grace.exceptions import ConfigError
 from grace.importer import find_all_importables, import_module
-
+from grace.model import Model
 
 ConfigReturn = Union[str, int, float, None]
 
@@ -43,13 +33,14 @@ class Application:
 
     def __init__(self) -> None:
         database_config_path: Path = Path("config/database.cfg")
-        
+
         if not database_config_path.exists():
             raise ConfigError("Unable to find the 'database.cfg' file.")
 
         self.__token: str = str(self.config.get("discord", "token"))
         self.__engine: Union[Engine, None] = None
 
+        self.environment: str = "development"
         self.command_sync: bool = True
         self.watch: bool = False
 
@@ -67,8 +58,9 @@ class Application:
         """Instantiate the session for querying the database."""
 
         if not self.__session:
-            session: sessionmaker = sessionmaker(bind=self.__engine)
-            self.__session = session()
+            # session_factory: sessionmaker = sessionmaker(bind=self.__engine)
+            # scoped_session_ = scoped_session(session_factory)
+            self.__session = Session(self.__engine)
 
         return self.__session
 
@@ -99,11 +91,11 @@ class Application:
     def database_infos(self) -> Dict[str, str]:
         return {
             "dialect": self.session.bind.dialect.name,
-            "database": self.session.bind.url.database
+            "database": self.session.bind.url.database,
         }
 
     @property
-    def database_exists(self):
+    def database_exists(self) -> bool:
         return database_exists(self.config.database_uri)
 
     def get_extension_module(self, extension_name) -> Union[str, None]:
@@ -134,9 +126,7 @@ class Application:
 
     def load_logs(self) -> None:
         file_handler: RotatingFileHandler = RotatingFileHandler(
-            f"logs/{self.config.current_environment}.log",
-            maxBytes=10000,
-            backupCount=5
+            f"logs/{self.config.current_environment}.log", maxBytes=10000, backupCount=5
         )
 
         basicConfig(
@@ -147,19 +137,24 @@ class Application:
 
         install(
             self.config.environment.get("log_level"),
-            fmt="".join([
-                "[%(asctime)s] %(programname)s %(funcName)s ",
-                "%(module)s %(levelname)s %(message)s"
-            ]),
+            fmt="".join(
+                [
+                    "[%(asctime)s] %(programname)s %(funcName)s ",
+                    "%(module)s %(levelname)s %(message)s",
+                ]
+            ),
             programname=self.config.current_environment,
         )
 
-    def load_database(self):
+    def load_database(self) -> None:
         """Loads and connects to the database using the loaded config"""
+
+        if not self.config.database_uri:
+            raise ValueError("No database uri.")
 
         self.__engine = create_engine(
             self.config.database_uri,
-            echo=self.config.environment.getboolean("sqlalchemy_echo")
+            echo=self.config.environment.getboolean("sqlalchemy_echo"),
         )
 
         if self.database_exists:
@@ -167,6 +162,8 @@ class Application:
                 self.__engine.connect()
             except OperationalError as e:
                 critical(f"Unable to load the 'database': {e}")
+
+        Model.set_engine(self.__engine)
 
     def unload_database(self):
         """Unloads the current database"""
@@ -176,7 +173,7 @@ class Application:
 
     def reload_database(self):
         """
-        Reload the database. This function can be use in case
+        Reload the database. This function can be used in case
         there's a dynamic environment change.
         """
 
@@ -198,11 +195,17 @@ class Application:
     def create_tables(self):
         """Creates all the tables for the current loaded database"""
 
+        if not self.__engine:
+            raise RuntimeError("Database engine is not initialized.")
+
         self.load_database()
         self.base.metadata.create_all(self.__engine)
 
     def drop_tables(self):
         """Drops all the tables for the current loaded database"""
+
+        if not self.__engine:
+            raise RuntimeError("Database engine is not initialized.")
 
         self.load_database()
         self.base.metadata.drop_all(self.__engine)
